Fixing a bug that was deleting all files in public on build
The webpack output attribute clean was set to true in the /public folder. This menat on every build all files were being deleted including fonts and images. Another folder has been created inside of the /public called /assets where the css and js will exist and the clean (delete) feature can be applied without impacting other assets. This is something that will need a better fix in future, this is only a quick fix to be able to continue dev